% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{initialize,PiMPDB-method}
\alias{initialize,PiMPDB-method}
\title{Constructor method of PiMPDB class}
\usage{
\S4method{initialize}{PiMPDB}(.Object, dbname = "character",
  dbuser = "character", dbpassword = "character", dbhost = "character",
  dbtype = "character", ...)
}
\arguments{
\item{dbname}{database name}

\item{dbuser}{database user name}

\item{dbpassword}{database password}

\item{dbhost}{database hostname}

\item{dbtype}{database type e.g. sqlite, mysql}

\item{.Object}{ An object:  see the Details section.}

\item{...}{data to include in the new object.  Named arguments
    correspond to slots in the class definition. Unnamed arguments must
    be objects from classes that this class extends.}
}
\value{
\code{PiMPDB} S4 object
}
\description{
Class \code{PiMPDB} defines a PiMPDB database connection object
}
\examples{
\dontrun{mydb1 <- new("PiMPDB", dbname="pimp.db", dbtype="sqlite")}
\dontrun{mydb2 <- new("PiMPDB", dbname="pimpdb", dbuser="pimp", dbpassword="mypassword", dbhost="localhost", dbtype="mysql")}
}

