% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{PiMPDB-class}
\alias{PiMPDB-class}
\title{An S4 class to represent a PiMPDB connection.}
\description{
An S4 class to represent a PiMPDB connection.
}
\section{Slots}{

\describe{
\item{\code{dbconnection}}{A DBMS specific DBIConnection e.g. RSQLite or RMySQL}
}}

